Add missing ApiExceptionFilterAttributeTests
